+++ .gitignore
@@ -4,20 +4,9 @@ public/files/
 # build output
 dist/
 .astro/
-.vercel/
 
 # dependencies
 node_modules/
-
-# logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-pnpm-debug.log*
-
-# environment variables
-.env
-.env.production
 
 # macOS-specific files
 .DS_Store
