+++ package.json
@@ -18,7 +18,6 @@
     "@tailwindcss/typography": "^0.5.15",
     "@types/react-dom": "^19.0.2",
     "@types/react": "^19.0.2",
-    "@vercel/og": "^0.6.4",
     "astro": "^5.1.1",
     "astro-icon": "^1.1.4",
     "astro-robots-txt": "^1.0.0",
@@ -27,6 +26,7 @@
     "react-dom": "^19.0.0",
     "react": "^19.0.0",
     "rehype-external-links": "^3.0.0",
+    "satori": "^0.12.0",
     "sharp-ico": "^0.1.5",
     "sharp": "^0.33.5",
     "tailwindcss": "^3.4.16"
