+++ src/pages/photography/metadata.json.ts
@@ -8,8 +8,10 @@ export async function GET(): Promise<Response> {
     data: {
       title: "Photography",
       subtitle: `by ${AUTHOR.name}`,
-      description:
-        "Enjoy my photography from around the world. You are free to use all the images for any purpose. When sharing, please credit me as the photographer. Thank you! 📸",
+      description: `Enjoy my photography from around the world.
+                    You are free to use all the images for any purpose.
+                    When sharing, please credit me as the photographer.
+                    Thank you! 📸`,
     },
     cover: photos[0],
   } satisfies Metadata;
